move web-related settings
